#5 refactor: global error 패키지를 exception으로 네이밍 변경
