+++ src/main/java/com/world/playstay/global/exception/GlobalHttpException.java
@@ -1,4 +1,4 @@
-package com.world.playstay.global.error;
+package com.world.playstay.global.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
