+++ src/main/java/com/world/playstay/global/exception/LogLevel.java
@@ -1,4 +1,4 @@
-package com.world.playstay.global.error;
+package com.world.playstay.global.exception;
 
 public enum LogLevel {
   TRACE, DEBUG, INFO, WARN, ERROR, FATAL, NONE
