+++ src/main/java/com/world/playstay/global/util/exception/EncryptException.java
@@ -1,7 +1,7 @@
 package com.world.playstay.global.util.exception;
 
-import com.world.playstay.global.error.GlobalHttpException;
-import com.world.playstay.global.error.LogLevel;
+import com.world.playstay.global.exception.GlobalHttpException;
+import com.world.playstay.global.exception.LogLevel;
 import org.springframework.http.HttpStatus;
 
 public class EncryptException extends GlobalHttpException {
