+++ src/main/java/com/world/playstay/user/controller/GuestController.java
@@ -1,6 +1,6 @@
 package com.world.playstay.user.controller;
 
-import com.world.playstay.global.error.GlobalExceptionResponse;
+import com.world.playstay.global.exception.GlobalExceptionResponse;
 import com.world.playstay.user.dto.request.CreateGuestRequest;
 import com.world.playstay.user.dto.response.GuestResponse;
 import com.world.playstay.user.entity.Guest;
