+++ src/main/java/com/world/playstay/user/exception/DuplicatedUserException.java
@@ -1,7 +1,7 @@
 package com.world.playstay.user.exception;
 
-import com.world.playstay.global.error.GlobalHttpException;
-import com.world.playstay.global.error.LogLevel;
+import com.world.playstay.global.exception.GlobalHttpException;
+import com.world.playstay.global.exception.LogLevel;
 import org.springframework.http.HttpStatus;
 
 public class DuplicatedUserException extends GlobalHttpException {
