+++ src/main/java/com/world/playstay/user/exception/UserNotFoundException.java
@@ -1,11 +1,12 @@
 package com.world.playstay.user.exception;
 
-import com.world.playstay.global.error.GlobalHttpException;
-import com.world.playstay.global.error.LogLevel;
+import com.world.playstay.global.exception.GlobalHttpException;
+import com.world.playstay.global.exception.LogLevel;
 import org.springframework.http.HttpStatus;
 
 public class UserNotFoundException extends GlobalHttpException {
-  public UserNotFoundException(String message){
+
+  public UserNotFoundException(String message) {
     super(HttpStatus.NOT_FOUND, message, LogLevel.ERROR);
   }
 }
